Try to use the high and low numbers for all intervals
OpenWeatherAPI provides 8 3-hour intervals for its 5 day forecast. So the attempt was to use the high and low from looking at all intervals between 6am and midnight. This API doesn't seem to be very accurate though. Maybe that's the free tier? Seems silly to offer inaccurate information though. The paid version shouldn't be to get better information.

So I'm going to ask Trae to refactor and allow other APIs to be used. We'll see how that goes.

Oh yeah. I'm trying out Trae for this app. Why not? I like the name. :)